+++ src/main/java/net/grid/vampiresdelight/common/mixin/plugin/VDMixinConfigPlugin.java
@@ -1,0 +1,53 @@
+package net.grid.vampiresdelight.common.mixin.plugin;
+
+import net.grid.vampiresdelight.common.utility.VDIntegrationUtils;
+import org.objectweb.asm.tree.ClassNode;
+import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
+import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
+
+import java.util.List;
+import java.util.Set;
+
+public class VDMixinConfigPlugin implements IMixinConfigPlugin {
+    @Override
+    public void onLoad(String mixinPackage) {
+
+    }
+
+    @Override
+    public String getRefMapperConfig() {
+        return null;
+    }
+
+    private static final List<String> APPLESKIN_MIXINS = List.of(
+            "net.grid.vampiresdelight.common.mixin.MixinFoodHelper"
+    );
+
+    @Override
+    public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
+        if (APPLESKIN_MIXINS.contains(mixinClassName) && !VDIntegrationUtils.isLoadingModPresent(VDIntegrationUtils.APPLESKIN))
+            return false;
+
+        return true;
+    }
+
+    @Override
+    public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {
+
+    }
+
+    @Override
+    public List<String> getMixins() {
+        return null;
+    }
+
+    @Override
+    public void preApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
+
+    }
+
+    @Override
+    public void postApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
+
+    }
+}