+++ src/main/java/net/grid/vampiresdelight/common/utility/VDTooltipUtils.java
@@ -17,29 +17,30 @@ public class VDTooltipUtils {
     public static void addFactionFoodToolTips(@NotNull List<Component> tooltip, @Nullable Player player, IPlayableFaction<?> foodFaction) {
         assert player != null;
 
-        if (!Helper.isVampire(player) && !VDConfiguration.HUNTER_TOOLTIPS_FOR_EVERYONE.get())
-            return;
+        // TODO: FIX BAT
+        if (!Helper.isVampire(player) && !VDConfiguration.HUNTER_TOOLTIPS_FOR_EVERYONE.get()) {
+            if (!Objects.equals(foodFaction, VReference.VAMPIRE_FACTION))
+                return;
+        }
 
         tooltip.add(Component.empty());
         tooltip.add(Component.translatable("text.vampirism.faction_specifics").withStyle(ChatFormatting.GRAY));
-        ChatFormatting color = ChatFormatting.GRAY;
+        ChatFormatting color;
 
-        if (foodFaction != null) {
-            if (Objects.equals(foodFaction, VReference.VAMPIRE_FACTION)) {
-                if (Helper.isVampire(player)) {
-                    color = ChatFormatting.DARK_GREEN;
-                } else {
-                    color = ChatFormatting.DARK_RED;
-                }
+        if (Objects.equals(foodFaction, VReference.VAMPIRE_FACTION)) {
+            if (Helper.isVampire(player)) {
+                color = ChatFormatting.DARK_GREEN;
             } else {
-                if (!Helper.isVampire(player)) {
-                    color = ChatFormatting.DARK_GREEN;
-                } else {
-                    color = ChatFormatting.DARK_RED;
-                }
+                color = ChatFormatting.DARK_RED;
             }
-
-            tooltip.add(Component.translatable(" ").append(foodFaction.getName()).withStyle(color));
+        } else {
+            if (!Helper.isVampire(player)) {
+                color = ChatFormatting.DARK_GREEN;
+            } else {
+                color = ChatFormatting.DARK_RED;
+            }
         }
+
+        tooltip.add(Component.literal(" ").append(foodFaction.getName()).withStyle(color));
     }
 }